refactor: Move utils
